requirements.txt
## !! Could not determine repository location
barniecopter==0.1.0
certifi==2020.12.5
chardet==4.0.0
idna==2.10
mergedeep==1.3.4
PyYAML==5.4.1
requests==2.25.1
responses==0.13.2
six==1.15.0
urllib3==1.26.4
